% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Bubble chart display of KEGG/GO results}
\usage{
bubble_plot(data, padjust = 0.05, top = 20, text.size = 4, color = "RdBu", ...)
}
\arguments{
\item{data}{R clusterprofiler package for KEGG or GO results}

\item{padjust}{such as padjust = 0.01, mean that p.adjust <= 0.01}

\item{top}{According to the order of P adjust value from small to large
the number of categories to show}

\item{text.size}{text size}

\item{color}{color see "RColorBrewer::display.brewer.all()"}

\item{...}{additional parameters}
}
\value{
bubble plot
}
\description{
Bubble chart display of KEGG/GO results
}
\examples{
\dontrun{
data(xfbdf)
library(clusterProfiler)
eg <- bitr(unique(xfbdf$target),
  fromType = "SYMBOL",
  toType = "ENTREZID",
  OrgDb = "org.Hs.eg.db"
)
KK <- enrichKEGG(
  gene = eg$ENTREZID,
  organism = "hsa",
  pvalueCutoff = 0.05
)
bubble_plot(KK)
}
}
