% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compound_data}
\alias{compound_data}
\title{compound_data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13 rows and 5 columns.
}
\usage{
compound_data
}
\description{
A data frame with 13 rows and 5 variables
}
\examples{
head(compound_data)

}
\keyword{datasets}
