% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compoundfinal}
\alias{compoundfinal}
\title{compoundfinal}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3397 rows and 5 columns.
}
\usage{
compoundfinal
}
\description{
A data frame with 3397rows and 5 variables
}
\examples{
head(compoundfinal)

}
\keyword{datasets}
