% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_plot.R
\name{degree_plot}
\alias{degree_plot}
\title{the degree of a network node}
\usage{
degree_plot(
  data,
  color = c("#A40545", "#7FCBA4", "#4B65AF"),
  top = 10,
  text.size = 12,
  text.width = 20,
  plot.set = "vertical",
  title = "the degree of networks",
  ...
)
}
\arguments{
\item{data}{data.frame
the column name must contain three columns:
herb, molecule, and target}

\item{color}{color, Recommended color:
colour1 <- c("#c0583d", "#daa152", "#b3c6dc")
colour2 <- c("#8fa1c7", "#ef916a", "#7dbfa5")
colour3 <- c("#7dbfa5", "#487eb3", "#d2342b")
colour4 <- c("#6176b5", "#DFC86B", "#87BD5B")
colour5 <- c("#852f88", "#eb990c", "#0f8096")
colour6 <- c("#3c4b8d", "#da2e20", "#3c884c")
colour7 <- c("#7aa354", "#d67d44", "#6594c4")
colour8 <- c("#501d8a", "#e55709", "#1c8041")
colour9 <- c("#A40545", "#7FCBA4", "#4B65AF")}

\item{top}{According to the order of P adjust value from small to large
the number of categories to show}

\item{text.size}{text size}

\item{text.width}{text width}

\item{plot.set}{"horizontal" or "vertical"}

\item{title}{title}

\item{...}{additional parameters}
}
\value{
bar plot
}
\description{
the degree of a network node
}
\examples{
degree_plot(xfbdf)
}
