% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_data}
\alias{disease_data}
\title{disease_data}
\format{
An object of class \code{data.frame} with 790 rows and 1 columns.
}
\source{
KEGG & GSEA
}
\usage{
disease_data
}
\description{
A character with 790 genes
}
\examples{
head(disease_data)

}
\keyword{datasets}
