% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_plot.R
\name{dot_plot}
\alias{dot_plot}
\title{Bubble graphs showed the results of GO and KEGG analysis}
\usage{
dot_plot(
  data,
  color = "RdBu",
  top = 15,
  text.size = 10,
  text.width = 35,
  title = NULL,
  ...
)
}
\arguments{
\item{data}{R clusterprofiler package for KEGG or GO results}

\item{color}{color see "RColorBrewer::display.brewer.all()"}

\item{top}{number of categories to show}

\item{text.size}{text size}

\item{text.width}{y-axis label length}

\item{title}{title}

\item{...}{additional parameters}
}
\value{
dotplot
}
\description{
Bubble graphs showed the results of GO and KEGG analysis
}
\examples{
\dontrun{
data(xfbdf, package = "TCMNP")
library(clusterProfiler)
library(org.Hs.eg.db)
library(DOSE)
eg <- bitr(unique(xfbdf$target),
  fromType = "SYMBOL",
  toType = "ENTREZID",
  OrgDb = "org.Hs.eg.db"
)
KK <- enrichKEGG(
  gene = eg$ENTREZID,
  organism = "hsa",
  pvalueCutoff = 0.05
)
KK <- setReadable(KK, "org.Hs.eg.db", keyType = "ENTREZID")
BP <- enrichGO(
  gene = eg$ENTREZID,
  "org.Hs.eg.db",
  ont = "BP",
  pvalueCutoff = 0.05,
  readable = TRUE
)
dot_plot(KK, title = "KEGG")
dot_plot(BP, title = "biological process")
}
}
