% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_sankey2.R
\name{dot_sankey2}
\alias{dot_sankey2}
\title{Sankey and dot graphs showed the results
after KEGG/GO screening of pathways and genes}
\usage{
dot_sankey2(
  data,
  sankey.text.size = 3,
  sankey.x.axis.text.position = 1,
  sankey.x.axis.text.size = 12,
  dot.color = "RdBu",
  dot.position = 8,
  dot.lable = "Description",
  dot.text.size = 12,
  dot.scale = 0.9,
  dot.x = 0.45,
  dot.y = 0.17,
  dot.width = 0.5,
  dot.height = 0.63,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{sankey.text.size}{Sankey diagram text size}

\item{sankey.x.axis.text.position}{Location of x-axis labels in sankey plot}

\item{sankey.x.axis.text.size}{Sankey diagram x-axis text size}

\item{dot.color}{dotplot color: see "RColorBrewer::display.brewer.all()"}

\item{dot.position}{dotplot position}

\item{dot.lable}{"ID" or "Description"}

\item{dot.text.size}{dotplot text size}

\item{dot.scale}{dotplot scale}

\item{dot.x}{dotplot horizontal position}

\item{dot.y}{dotplot up and down position}

\item{dot.width}{dotplot width}

\item{dot.height}{dotplot height}

\item{...}{additional parameters}
}
\value{
figure
}
\description{
Sankey and dot graphs showed the results
after KEGG/GO screening of pathways and genes
}
\examples{
\dontrun{
data(kegg.filter, package = "TCMNP")
dot_sankey2 (kegg.filter)
}
}
