% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etcm.R
\name{etcm}
\alias{etcm}
\title{ETCM database download data collation}
\usage{
etcm(data, herb = NULL)
}
\arguments{
\item{data}{data downloaded from ETCM database}

\item{herb}{herb herb name}
}
\value{
data.frame
}
\description{
ETCM database download data collation
}
\examples{
data("mahuang", package = "TCMNP")
data <- etcm(mahuang, herb = "ma huang")
head(data)
}
