% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_barplot.R
\name{go_barplot}
\alias{go_barplot}
\title{Bar graphs showed the results of GO results}
\usage{
go_barplot(
  go.diff,
  top = 5,
  plot.set = "vertical",
  facet = TRUE,
  color = c("#A40545", "#7FCBA4", "#4B65AF"),
  bar.width = 0.8,
  text.size = 12,
  text.width = 30,
  title = "GO enrichment of cluster",
  ...
)
}
\arguments{
\item{go.diff}{R clusterprofiler package for enrichGO results}

\item{top}{According to the order of P adjust value from small to large
the number of categories to show}

\item{plot.set}{"horizontal" or "vertical"}

\item{facet}{facet}

\item{color}{color, Recommended color:
colour1 <- c("#c0583d", "#daa152", "#b3c6dc")
colour2 <- c("#8fa1c7", "#ef916a", "#7dbfa5")
colour3 <- c("#7dbfa5", "#487eb3", "#d2342b")
colour4 <- c("#6176b5", "#DFC86B", "#87BD5B")
colour5 <- c("#852f88", "#eb990c", "#0f8096")
colour6 <- c("#3c4b8d", "#da2e20", "#3c884c")
colour7 <- c("#7aa354", "#d67d44", "#6594c4")
colour8 <- c("#501d8a", "#e55709", "#1c8041")
colour9 <- c("#A40545", "#7FCBA4", "#4B65AF")}

\item{bar.width}{bar width}

\item{text.size}{text size}

\item{text.width}{text width}

\item{title}{title}

\item{...}{additional parameters}
}
\value{
bar plot
}
\description{
Bar graphs showed the results of GO results
}
\examples{
\dontrun{
library(clusterProfiler)
library(org.Hs.eg.db)
library(DOSE)
eg <- bitr(unique(xfbdf$target),
  fromType = "SYMBOL",
  toType = "ENTREZID",
  OrgDb = "org.Hs.eg.db"
)
go.diff <- enrichGO(
  gene = eg$ENTREZID,
  org.Hs.eg.db,
  pAdjustMethod = "BH",
  pvalueCutoff = 0.01,
  qvalueCutoff = 0.05,
  ont = "all",
  readable = T
)
go_barplot(go.diff)
}
}
