% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_cir.R
\name{go_cir}
\alias{go_cir}
\title{Chord diagram display of GO results}
\usage{
go_cir(
  go.diff,
  top = 5,
  id.color = c("#852f88", "#eb990c", "#0f8096"),
  id.cex = 0.8,
  gene.color = "#CFB0D4",
  gene.cex = 0.8,
  ratio.color = c("#6291bd", "#B1D1E7"),
  ratio.cex = 0.8,
  p.color = c("#eda9aa", "#c25254"),
  p.cex = 0.8,
  richFactor.cex = 0.7,
  legend.position = c(0.8, 0.45),
  text.width = 35,
  ...
)
}
\arguments{
\item{go.diff}{R clusterprofiler package for enrichGO results}

\item{top}{According to the order of P adjust value from small to large
the number of categories to show}

\item{id.color}{The color of the first layer ID of the path
Recommended color:
colour1 <- c("#c0583d", "#daa152", "#b3c6dc")
colour2 <- c("#8fa1c7", "#ef916a", "#7dbfa5")
colour3 <- c("#7dbfa5", "#487eb3", "#d2342b")
colour4 <- c("#6176b5", "#DFC86B", "#87BD5B")
colour5 <- c("#852f88", "#eb990c", "#0f8096")
colour6 <- c("#3c4b8d", "#da2e20", "#3c884c")
colour7 <- c("#7aa354", "#d67d44", "#6594c4")
colour8 <- c("#501d8a", "#e55709", "#1c8041")
colour9 <- c("#A40545", "#7FCBA4", "#4B65AF")}

\item{id.cex}{The text size of the first layer ID of the path}

\item{gene.color}{The color of the number of all genes in this pathway}

\item{gene.cex}{The font size of the number of all genes in this pathway}

\item{ratio.color}{（1）The color of the number of genes enriched this time on this pathway
（2）Other genes involved in this enrichment analysis}

\item{ratio.cex}{The text size of the numbers involved in this enrichment analysis}

\item{p.color}{p adjust color}

\item{p.cex}{The text size of padjust}

\item{richFactor.cex}{The text size of richFactor}

\item{legend.position}{legend position}

\item{text.width}{legend text width}

\item{...}{additional parameters}
}
\value{
chord diagram
}
\description{
Chord diagram display of GO results
}
\examples{
\dontrun{
library(clusterProfiler)
library(org.Hs.eg.db)
library(DOSE)
data(xfbdf, package = "TCMNP")
eg <- bitr(unique(xfbdf$target),
  fromType = "SYMBOL",
  toType = "ENTREZID",
  OrgDb = "org.Hs.eg.db"
)
go.diff <- enrichGO(
  gene = eg$ENTREZID,
  org.Hs.eg.db,
  pAdjustMethod = "BH",
  pvalueCutoff = 0.01,
  qvalueCutoff = 0.05,
  ont = "all",
  readable = T
)
go_cir(go.diff)
}
}
