% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_lollipop.R
\name{go_lollipop}
\alias{go_lollipop}
\title{Lollipop chart showed the results of GO results}
\usage{
go_lollipop(
  go.diff,
  top = 5,
  color = "RdBu",
  text.size = 12,
  text.width = 35,
  title = "GO enrichment of cluster",
  ...
)
}
\arguments{
\item{go.diff}{R clusterprofiler package for enrichGO results}

\item{top}{According to the order of p adjust value from small to large
the number of categories to show}

\item{color}{color see "RColorBrewer::display.brewer.all()"}

\item{text.size}{text size}

\item{text.width}{text width}

\item{title}{title}

\item{...}{additional parameters}
}
\value{
lollipop plot
}
\description{
Lollipop chart showed the results of GO results
}
\examples{
\dontrun{
library(clusterProfiler)
library(org.Hs.eg.db)
library(DOSE)
data(xfbdf, package = "TCMNP")
eg <- bitr(unique(xfbdf$target),
  fromType = "SYMBOL",
  toType = "ENTREZID",
  OrgDb = "org.Hs.eg.db"
)
go.diff <- enrichGO(
  gene = eg$ENTREZID,
  org.Hs.eg.db,
  pAdjustMethod = "BH",
  pvalueCutoff = 0.01,
  qvalueCutoff = 0.05,
  ont = "all",
  readable = T
)
go_lollipop(go.diff)
}
}
