% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{herb_pm}
\alias{herb_pm}
\title{Herbal Properties and Meridians}
\format{
An object of class \code{data.frame} with 409 rows and 5 columns.
}
\source{
\url{https://db.ouryao.com/yd2020/index.php?cid=1}
}
\usage{
herb_pm
}
\description{
Chinese Pharmacopoeia commonly used herbs property and flavor, tropism
The datasets was from Chinese Pharmacopoeia
A data frame with 409 rows and 5 variables
}
\examples{
head(herb_pm)
}
\keyword{datasets}
