% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_target.R
\name{herb_target}
\alias{herb_target}
\title{Screen targets based on ingredients from Chinese herbal medicine}
\usage{
herb_target(herb, OB.value = 30, DL.value = 0.18, type = "Herb_cn_name", ...)
}
\arguments{
\item{herb}{The name of the compound Chinese herbal medicine}

\item{OB.value}{oral bioavailability}

\item{DL.value}{drug-likeness}

\item{type}{Writing format of herb name:
"Herb_name_pin_yin" or "Herb_cn_name"}

\item{...}{additional parameters}
}
\value{
data.frame
}
\description{
Screen targets based on ingredients from Chinese herbal medicine
}
\examples{
herbs <- c("ma huang", "ku xing ren")
fufang <- herb_target(herbs, type = "Herb_name_pin_yin")
head(fufang)
}
