% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_gsea}
\alias{human_gsea}
\title{human_gsea}
\format{
An object of class \code{data.frame} with 7321 rows and 2 columns.
}
\usage{
human_gsea
}
\description{
A data frame with 7189 rows and 2 variables
}
\examples{
head(human_gsea)

}
\keyword{datasets}
