% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_kegg}
\alias{human_kegg}
\title{human_kegg}
\format{
An object of class \code{data.frame} with 35539 rows and 2 columns.
}
\usage{
human_kegg
}
\description{
A data frame with 36788 rows and 2 variables
}
\examples{
head(human_kegg)

}
\keyword{datasets}
