% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg.filter}
\alias{kegg.filter}
\title{kegg.filter}
\format{
An object of class \code{data.frame} with 8 rows and 9 columns.
}
\source{
example data from the KEGG results
}
\usage{
kegg.filter
}
\description{
The results after KEGG/GO screening of pathways and genes
A data frame with 8 rows and 9 variables
}
\examples{
head(kegg.filter)

}
\keyword{datasets}
