% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KK}
\alias{KK}
\title{KK}
\format{
An object of class \code{enrichResult} with 193 rows and 9 columns.
}
\source{
R clusterprofiler package for KEGG results
}
\usage{
KK
}
\description{
The results of R clusterprofiler package for KEGG results
S4 object of class enrich results
}
\examples{
head(KK@result)

}
\keyword{datasets}
