% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipop_plot.R
\name{lollipop_plot}
\alias{lollipop_plot}
\title{Lollipop chart display of KEGG/GO results}
\usage{
lollipop_plot(
  data,
  top = 15,
  color = "RdBu",
  title = NULL,
  text.size = 10,
  text.width = 35,
  ...
)
}
\arguments{
\item{data}{R clusterprofiler package for KEGG and GO results}

\item{top}{According to the order of p adjust value from small to large
the number of categories to show}

\item{color}{color color see "RColorBrewer::display.brewer.all()"}

\item{title}{title}

\item{text.size}{text size}

\item{text.width}{text width}

\item{...}{additional parameters}
}
\value{
lollipop plot
}
\description{
Lollipop chart display of KEGG/GO results
}
\examples{
\dontrun{
data(xfbdf)
library(org.Hs.eg.db)
library(clusterProfiler)
eg <- bitr(unique(xfbdf$target),
  fromType = "SYMBOL",
  toType = "ENTREZID",
  OrgDb = "org.Hs.eg.db"
)
KK <- enrichKEGG(
  gene = eg$ENTREZID,
  organism = "hsa",
  pvalueCutoff = 0.05
)
MF<- enrichGO(
  gene = eg$ENTREZID,
  "org.Hs.eg.db",
  ont = "MF",
  pvalueCutoff = 0.05,
  readable = TRUE
)
lollipop_plot(MF, title = "Molecular Function", color = "Spectral")
lollipop_plot(KK)
}
}
