% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mahuang}
\alias{mahuang}
\title{ma huang}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 28 rows and 2 columns.
}
\source{
\url{http://www.tcmip.cn/ETCM/index.php/Home/Index/yc_details.html?id=212}
}
\usage{
mahuang
}
\description{
The datasets was from ETCM database.
A data frame with 28 rows and 2 variables
}
\examples{
head(mahuang)

}
\keyword{datasets}
