% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu_gsea}
\alias{mu_gsea}
\title{mu_gsea}
\format{
An object of class \code{data.frame} with 7189 rows and 2 columns.
}
\usage{
mu_gsea
}
\description{
A data frame with 7189 rows and 2 variables
}
\examples{
head(mu_gsea)

}
\keyword{datasets}
