% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu_kegg}
\alias{mu_kegg}
\title{mu_kegg}
\format{
An object of class \code{data.frame} with 36788 rows and 2 columns.
}
\usage{
mu_kegg
}
\description{
A data frame with 36788 rows and 2 variables
}
\examples{
head(compound_data)

}
\keyword{datasets}
