% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_data}
\alias{network_data}
\title{network_data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 3 columns.
}
\source{
network_data
}
\usage{
network_data
}
\description{
A data frame with 100rows and 3 variables
}
\examples{
head(network_data)

}
\keyword{datasets}
