% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_ccplot.R
\name{pathway_ccplot}
\alias{pathway_ccplot}
\title{circle chart showed the results of GO and KEGG analysis}
\usage{
pathway_ccplot(
  data,
  top = 5,
  label.name = "Description",
  root = NULL,
  color.node = "Paired",
  color.alpha = 0.5,
  text.size = c(3, 4),
  ...
)
}
\arguments{
\item{data}{R clusterprofiler package for KEGG and GO results
The data format must be S4 object with KEGG and GO results or
a data frame with KEGG and GO results or
a data frame with two columns.}

\item{top}{According to the order of p adjust value from small to large
the number of categories to show}

\item{label.name}{"ID" or "Description"}

\item{root}{root name}

\item{color.node}{node color}

\item{color.alpha}{color alpha}

\item{text.size}{text size
(1) first circle text size
(2) Second circle text size}

\item{...}{additional parameters}
}
\value{
figure
}
\description{
circle chart showed the results of GO and KEGG analysis
}
\examples{
\dontrun{
data("KK", package = "TCMNP")
library(dplyr)
pathway_ccplot(KK, root = "KEGG")
}

}
