% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_cirplot.R
\name{pathway_cirplot}
\alias{pathway_cirplot}
\title{Display the relationship between genes and pathways
enriched by KEGG or GO in the form of a chord diagram}
\usage{
pathway_cirplot(
  data,
  top = 5,
  start.degree = -90,
  label.name = "Description",
  text.width = 15,
  text.size = 0.6,
  color = "RdBu",
  ...
)
}
\arguments{
\item{data}{R clusterprofiler package for KEGG and GO results
The data format must be S4 object with KEGG and GO results or
a data frame with KEGG and GO results or
a data frame with two columns.}

\item{top}{According to the order of p adjust value from small to large
the number of categories to show}

\item{start.degree}{the angle at which the circle plot starts}

\item{label.name}{"ID" or "Description"}

\item{text.width}{label length}

\item{text.size}{the font size of each layer circle}

\item{color}{color see "RColorBrewer::display.brewer.all()"}

\item{...}{additional parameters}
}
\value{
chord diagram
}
\description{
Display the relationship between genes and pathways
enriched by KEGG or GO in the form of a chord diagram
}
\examples{
\dontrun{
data(xfbdf, package = "TCMNP")
library(org.Hs.eg.db)
library(clusterProfiler)
library(DOSE)
library(tidyverse)
eg <- bitr(unique(xfbdf$target),
  fromType = "SYMBOL",
  toType = "ENTREZID",
  OrgDb = "org.Hs.eg.db"
)
KK <- enrichKEGG(
  gene = eg$ENTREZID,
  organism = "hsa",
  pvalueCutoff = 0.05
)
KK <- setReadable(KK, "org.Hs.eg.db", keyType = "ENTREZID")
pathway_cirplot(KK)
}
}
