% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppi_data}
\alias{ppi_data}
\title{ppi_data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 3 columns.
}
\source{
\url{https://string-db.org/}
}
\usage{
ppi_data
}
\description{
Protein-protein interaction analysis was performed using STRING database.
A data frame with 2234 rows and 3 variables
}
\examples{
ppi_data<- data(ppi_data)
head(ppi_data)

}
\keyword{datasets}
