% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_plot.R
\name{ppi_plot}
\alias{ppi_plot}
\title{Protein-protein interaction analysis}
\usage{
ppi_plot(
  data,
  node.color = "RdBu",
  node.size = c(1, 10),
  label.size = 4,
  label.degree = 5,
  label.repel = TRUE,
  edge.color = "lightgrey",
  edge.width = c(0.2, 2),
  rem.dis.inter = FALSE,
  graph.layout = "kk"
)
}
\arguments{
\item{data}{Human protein-protein interaction (PPI)
data were downloaded from the STRING database.}

\item{node.color}{node color: see "RColorBrewer::display.brewer.all()"}

\item{node.size}{node size}

\item{label.size}{markup text size}

\item{label.degree}{The node degree is the number of connections that
the node has with the other nodes.
Nodes with connections greater than or
equal to degree will be displayed.}

\item{label.repel}{label repel}

\item{edge.color}{edge color}

\item{edge.width}{edge width}

\item{rem.dis.inter}{remove single free unconnected nodes}

\item{graph.layout}{Network Diagram Layout:
"kk", "nicely", "sphere", "circle",
"bipartite", "star","tree", "randomly",
"gem", "graphopt","lgl", "grid",
"mds", "sugiyama"}
}
\value{
network diagrams
}
\description{
Protein-protein interaction analysis
}
\examples{
\dontrun{
data(ppi_data, package = "TCMNP")
ppi_plot(ppi_data, label.repel = FALSE)
}
}
