% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sankey_data}
\alias{sankey_data}
\title{sankey_data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 30 rows and 4 columns.
}
\source{
xfbdf
}
\usage{
sankey_data
}
\description{
A data frame with 100 rows and 4 variables
}
\examples{
head(sankey_data)

}
\keyword{datasets}
