% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_herb.R
\name{target_herb}
\alias{target_herb}
\title{Find Chinese medicines and their active compounds through the target}
\usage{
target_herb(gene, OB.value = 30, DL.value = 0.18, ...)
}
\arguments{
\item{gene}{a vector of gene name}

\item{OB.value}{oral bioavailability}

\item{DL.value}{drug-likeness}

\item{...}{additional parameters}
}
\value{
data.frame
}
\description{
Find Chinese medicines and their active compounds through the target
}
\examples{
gene <- c("MAPK1", "JUN", "FOS", "RAC1", "IL1", "IL6")
target_herbs <- target_herb(gene)
}
