% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_serch.R
\name{target_serch}
\alias{target_serch}
\title{Screen target based on ingredients from Chinese herbal medicine}
\usage{
target_serch(target, ...)
}
\arguments{
\item{target}{The name of the target filter}

\item{...}{additional parameters}
}
\value{
data.frame
}
\description{
Screen target based on ingredients from Chinese herbal medicine
}
\examples{
target <- c("PRSS3", "ADH1C")
target_filter<- target_serch(target, type = "Herb_name_pin_yin")
head(target_filter)
}
