% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_alluvial.R
\name{tcm_alluvial}
\alias{tcm_alluvial}
\title{Alluvial diagram}
\usage{
tcm_alluvial(
  data,
  text.size = 3,
  text.position = 0,
  axis.text.x.size = 12,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{text.size}{text size}

\item{text.position}{text position 0, 0.5, 1}

\item{axis.text.x.size}{Location of x-axis labels in sankey plot}

\item{...}{additional parameters}
}
\value{
figure
}
\description{
Alluvial diagram
}
\examples{
data("xfbdf", package = "TCMNP")
data <- xfbdf \%>\% dplyr::sample_n(30, replace = FALSE)
tcm_alluvial(data, text.position = 1)
}
