% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_comp.R
\name{tcm_comp}
\alias{tcm_comp}
\title{Drug composition in prescription}
\usage{
tcm_comp(
  data,
  start.degree = 180,
  color = "Paired",
  radius.ratio = c(0.4, 0.2, 0.2),
  text.size = c(0.9, 0.9, 0.9)
)
}
\arguments{
\item{data}{data must consist of herb and weight columns}

\item{start.degree}{the angle at which the circle plot starts}

\item{color}{the color of circle plot.
see "RColorBrewer::display.brewer.all()"}

\item{radius.ratio}{from the outside to the inside,
the size of the proportion of each layer of the circle plot}

\item{text.size}{the font size of each layer circle}
}
\value{
figure
}
\description{
Drug composition in prescription
}
\examples{
\dontrun{
xfbdf.compostion <- data.frame(
  herb = c(
    "mahuang", "kuxingren", "shengshigao",
    "shengyiren", "maocangzhu", "guanghuoxiang",
    "qinghaocao", "mabiancao", "ganlugen", "tinglizi",
    "huajuhong", "shenggancao", "huzhang"
  ),
  weight = c(6, 15, 30, 30, 10, 15, 12, 30, 30, 15, 15, 10, 20)
)
tcm_comp(xfbdf.compostion)
}
}
