% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_compound.R
\name{tcm_compound}
\alias{tcm_compound}
\title{Drug composition in prescription}
\usage{
tcm_compound(
  data,
  start.degree = 180,
  first.color = "Spectral",
  second.color = "RdBu",
  third.color = c("#217AB3", "#A6CEE3", "#5EB9A8", "#33A02C", "#FFFF99", "#FDBF6F",
    "#FB9A99", "#EF5B5B", "#E31A1C", "lightgrey"),
  fourth.color = "Spectral",
  fifth.color = "Set2",
  radius.ratio = c(0.125, 0.125, 0.125, 0.125, 0.125, 0.125),
  text.size = c(0.8, 0.8, 0.8, 0.8, 0.8, 0.8)
)
}
\arguments{
\item{data}{data must consist of herb and weight columns}

\item{start.degree}{the angle at which the circle plot starts}

\item{first.color}{first circle color}

\item{second.color}{second circle color}

\item{third.color}{third circle color}

\item{fourth.color}{fouth circle color}

\item{fifth.color}{fifth circle color
see "RColorBrewer::display.brewer.all()"}

\item{radius.ratio}{from the outside to the inside,
the size of the proportion of each layer of the circle plot}

\item{text.size}{the font size of each layer circle}
}
\value{
figure
}
\description{
Drug composition in prescription
}
\examples{
\dontrun{
herb = c("ma huang", "ku xing ren", "hua ju hong",
           "cang zhu", "guang huo xiang", "yi yi ren",
           "hu zhang", "qing hao", "ma bian cao", "lu gen",
           "ting li zi", "shi gao", "gan cao")
xfbdf.comp = data.frame( herb = herb,
  weight = c(6, 15, 15, 10, 15, 30, 20, 12, 30, 30, 15, 30, 10),
  property = herb_pm[match(herb, herb_pm$Herb_name_pinyin), ]$Property,
  flavor = herb_pm[match(herb, herb_pm$Herb_name_pinyin), ]$Flavor,
  meridian = herb_pm[match(herb, herb_pm$Herb_name_pinyin), ]$Meridian
)
tcm_compound(xfbdf.comp)
}
}
