% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_net.R
\name{tcm_net}
\alias{tcm_net}
\title{Network interaction diagram of herbs, ingredients, and targets}
\usage{
tcm_net(
  network.data,
  node.color = "RdBu",
  node.size = c(2, 8),
  label.size = 4,
  label.degree = 3,
  label.repel = TRUE,
  edge.color = "lightgrey",
  edge.width = c(0.2, 2),
  graph.layout = "kk",
  rem.dis.inter = FALSE
)
}
\arguments{
\item{network.data}{data frame
must contain herb, molecule, target three columns of data}

\item{node.color}{node color
see "RColorBrewer::display.brewer.all()"}

\item{node.size}{node size}

\item{label.size}{label size}

\item{label.degree}{the node degree is the number of connections that
the node has with the other nodes.
Nodes with connections greater than or
equal to degree will be displayed.}

\item{label.repel}{label repel}

\item{edge.color}{edge color}

\item{edge.width}{edge width}

\item{graph.layout}{Network Diagram Layout:
"kk", "nicely", "circle", "sphere",
"bipartite", "star", "tree", "randomly",
"gem", "graphopt","lgl", "grid",
"mds", "sugiyama","fr"}

\item{rem.dis.inter}{remove single free unconnected nodes}
}
\value{
Network Diagram
}
\description{
Network interaction diagram of herbs, ingredients, and targets
}
\examples{
\dontrun{
data("xfbdf", package = "TCMNP")
network.data <- xfbdf \%>\%
  dplyr::select(herb, molecule, target) \%>\%
  sample_n(100, replace = FALSE) \%>\%
  as.data.frame()
tcm_net(network.data, node.color= "Spectral",
label.degree = 1, rem.dis.inter = TRUE,graph.layout = "fr",
label.size = 3)
}
}
