% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_pre_plot.R
\name{tcm_pre_plot}
\alias{tcm_pre_plot}
\title{Bar graphs showed the results of tcm_prescription  analysis}
\usage{
tcm_pre_plot(
  data,
  top = 10,
  color = "RdBu",
  text.size = 10,
  text.bar = 4,
  text.width = 35,
  title = NULL,
  ...
)
}
\arguments{
\item{data}{tcm_prescription  results}

\item{top}{according to the order of p adjust value from small to large
the number of categories to show}

\item{color}{color see "RColorBrewer::display.brewer.all()"}

\item{text.size}{text size}

\item{text.bar}{The size of the text at the top of the bar graph}

\item{text.width}{y-axis label length}

\item{title}{title}

\item{...}{additional parameters}
}
\value{
barplot
}
\description{
Bar graphs showed the results of tcm_prescription  analysis
}
\examples{
\dontrun{
data(disease_data, package = "TCMNP")
compound<- tcm_prescription(disease_data)[[2]]
tcm_pre_plot(compound)
}
}
