% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_prescription.R
\name{tcm_prescription}
\alias{tcm_prescription}
\title{Screen herbs and compound based on disease gene}
\usage{
tcm_prescription(
  data,
  compound.source = "tcmp",
  OB.value = 30,
  DL.value = 0.18,
  freq.value = 50,
  digits.value = 2,
  ...
)
}
\arguments{
\item{data}{disease gene}

\item{compound.source}{Writing format of herb name:
Chinese medicine prescription(tcmp) and Chinese patent medicine(cpm)
"tcmp" or "pcm"}

\item{OB.value}{oral bioavailability}

\item{DL.value}{drug-likeness}

\item{freq.value}{herbs freqence}

\item{digits.value}{Keep the number of decimal places}

\item{...}{additional parameters}
}
\value{
data.frame
}
\description{
Screen herbs and compound based on disease gene
}
\examples{
data(disease_data, package = "TCMNP")
herbs <- tcm_prescription(disease_data)[[1]]
compound <- tcm_prescription(disease_data)[[2]]
}
