% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_sankey.R
\name{tcm_sankey}
\alias{tcm_sankey}
\title{Sankey diagram}
\usage{
tcm_sankey(
  data_sankey,
  text.size = 3,
  text.position = 0,
  x.axis.text.size = 12,
  ...
)
}
\arguments{
\item{data_sankey}{data.frame}

\item{text.size}{text size}

\item{text.position}{text position 0, 0.5, 1}

\item{x.axis.text.size}{Location of x-axis labels in sankey plot}

\item{...}{additional parameters}
}
\value{
figure
}
\description{
Sankey diagram
}
\examples{
data("xfbdf", package = "TCMNP")
data <- xfbdf \%>\% dplyr::sample_n(30, replace = FALSE)
tcm_sankey(data, text.position = 1)
}
