% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm_sankey_dot.R
\name{tcm_sankey_dot}
\alias{tcm_sankey_dot}
\title{Herb-ingredient-target-function enrichment analysis}
\usage{
tcm_sankey_dot(
  data,
  sankey.text.size = 3,
  sankey.x.axis.text.position = 1,
  sankey.x.axis.text.size = 12,
  molecule.lable = "molecule_id",
  dot.color = "RdBu",
  dot.position = 6,
  dot.lable = "Description",
  dot.text.size = 12,
  dot.scale = 0.72,
  dot.x = 0.59,
  dot.y = -0.173,
  dot.width = 0.48,
  dot.height = 1.33,
  ...
)
}
\arguments{
\item{data}{data.frame}

\item{sankey.text.size}{Sankey diagram text size}

\item{sankey.x.axis.text.position}{Location of x-axis labels in sankey plot}

\item{sankey.x.axis.text.size}{Sankey diagram x-axis text size}

\item{molecule.lable}{"molecule" or ""molecule_id"}

\item{dot.color}{dotplot color: see "RColorBrewer::display.brewer.all()"}

\item{dot.position}{dotplot position}

\item{dot.lable}{"ID" or "Description"}

\item{dot.text.size}{dotplot text size}

\item{dot.scale}{dotplot scale}

\item{dot.x}{dotplot horizontal position}

\item{dot.y}{dotplot up and down position}

\item{dot.width}{dotplot width}

\item{dot.height}{dotplot height}

\item{...}{additional parameters}
}
\value{
figure
}
\description{
Herb-ingredient-target-function enrichment analysis
}
\examples{
\dontrun{
data(xfbdf, package = "TCMNP")
library(clusterProfiler)
library(org.Hs.eg.db)
library(DOSE)
eg <- bitr(unique(xfbdf$target),
  fromType = "SYMBOL",
  toType = "ENTREZID",
  OrgDb = "org.Hs.eg.db"
)
KK <- enrichKEGG(
  gene = eg$ENTREZID,
  organism = "hsa",
  pvalueCutoff = 0.05
)
KK <- setReadable(KK, "org.Hs.eg.db", keyType = "ENTREZID") \%>\%
  dplyr::mutate(richFactor = Count / as.numeric(sub("/\\\\d+", "", BgRatio)))
path <- separate_rows(KK@result, geneID, sep = "/")
data_sankey <- left_join(xfbdf, path,
  by = c("target" = "geneID"),
  relationship = "many-to-many"
) \%>\%
  dplyr::distinct() \%>\%
  tidyr::drop_na() \%>\%
  dplyr::sample_n(30, replace = FALSE) \%>\%
  as.data.frame()
tcm_sankey_dot(data_sankey)
}
}
