% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCMNP-package.R
\docType{package}
\name{TCMNP-package}
\alias{TCMNP}
\alias{TCMNP-package}
\title{TCMNP: Traditional Chinese Medicine Network Pharmacology}
\description{
A visualization tool for network pharmacology data analysis.
}
\author{
\strong{Maintainer}: Jinkun Liu \email{liujinkun007@gmail.com} (\href{https://orcid.org/0000-0003-2212-9617}{ORCID})

}
\keyword{internal}
