% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_cirplot.R
\name{tf_cirplot}
\alias{tf_cirplot}
\title{Display the relationship between transcription factors and target genes}
\usage{
tf_cirplot(data, start.degree = 270, text.size = 0.6, color = "RdBu", ...)
}
\arguments{
\item{data}{data.frame}

\item{start.degree}{the angle at which the circle plot starts}

\item{text.size}{the font size of each layer circle}

\item{color}{color see "RColorBrewer::display.brewer.all()"}

\item{...}{additional parameters}
}
\value{
chord diagram
}
\description{
Display the relationship between transcription factors and target genes
}
\examples{
\dontrun{
data(xfbdf, package = "TCMNP")
tf_data <- tf_filter(xfbdf$target)
set.seed(1234)
data <- tf_data[, 1:2] \%>\%
  distinct() \%>\%
  sample_n(100)
tf_cirplot(data, color = "Spectral")
}
}
