% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tf_data}
\alias{tf_data}
\title{tf_data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 2 columns.
}
\source{
\url{https://www.grnpedia.org/trrust/}
}
\usage{
tf_data
}
\description{
The datasets was from TRRUST database
A data frame with 100 rows and 2 variables
}
\examples{
head(tf_data)
}
\keyword{datasets}
