% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_filter.R
\name{tf_filter}
\alias{tf_filter}
\title{Finding transcription factors and their target genes}
\usage{
tf_filter(data)
}
\arguments{
\item{data}{The data must be character as human or mouse gene symbol}
}
\value{
data.frame
}
\description{
Finding transcription factors and their target genes
}
\examples{
data(xfbdf, package = "TCMNP")
newdata <- tf_filter(xfbdf$target)
head(newdata)
}
