% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trrust}
\alias{trrust}
\title{trrust}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16453 rows and 6 columns.
}
\usage{
trrust
}
\description{
A data frame with 7189 rows and 2 variables
}
\examples{
head(trrust)

}
\keyword{datasets}
