% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_net.R
\name{venn_net}
\alias{venn_net}
\title{venn network}
\usage{
venn_net(
  data,
  node.color = "Spectral",
  node.size = c(1, 10),
  label.size = 3,
  label.degree = 1,
  label.repel = TRUE,
  edge.color = c("#5BAA56", "#f49d98", "#4186B7", "#8679BE"),
  edge.alpha = 0.5,
  edge.width = 0.5,
  edge.type = "arc",
  graph.layout = "kk"
)
}
\arguments{
\item{data}{data frame}

\item{node.color}{node color: see "RColorBrewer::display.brewer.all()"}

\item{node.size}{node size}

\item{label.size}{markup text size}

\item{label.degree}{The node degree is the number of connections that
the node has with the other nodes.
Nodes with connections greater than or
equal to degree will be displayed.}

\item{label.repel}{label repel: "TRUE" or "FALSE"}

\item{edge.color}{edge color
venn.color1 <- c("#b1d1e7", "#f0a4af", "#e5dfae", "#a8e1da", "#f8d068")
venn.color2 <- c("#d3e2b7", "#fbca50", "#c9d5d4", "#baa28a", "#4caecc")
venn.color3 <- c("#C1BCBF", "#B1D0A9", "#FFCCC3", "#FFD5AB", "#BCCBE5")
venn.color4 <- c("#e4c9b2", "#92C2DD", "#f49d98", "#fcd68f", "#629076")
venn.color5 <- c("#FF8748", "#5BAA56", "#B8BB5B", "#4186B7", "#8679BE")
venn.color6 <- c("#e6c09e", "#0d5888", "#cb8b3e", "#9cd6d6", "#dbcb09")
venn.color7 <- c("#E41A1C", "#1E90FF", "#FF8C00", "#4DAF4A", "#75cbdc")
venn.color8 <- c("#604192", "#139177", "#ed9e08", "#f56f08", "#4caecc")
venn.color9 <- c("#d37a20", "#dbcb09", "#3a9cbc", "#dd7208", "#a30019")}

\item{edge.alpha}{edge alpha}

\item{edge.width}{edge width}

\item{edge.type}{"arc", "fan", "hive"}

\item{graph.layout}{Network Diagram Layout:
"kk", "nicely", "sphere", "circle",
"bipartite", "star","tree", "randomly",
"gem", "graphopt","lgl", "grid",
"mds", "sugiyama","fr"}
}
\value{
venn network diagrams
}
\description{
venn network
}
\examples{
\dontrun{
data(venn_data, package = "TCMNP")
set.seed(1234)
gene <- names(sort(table(venn_data$gene), decreasing = TRUE))[1:50]
data <- venn_data[venn.data$gene \%in\% gene, ]
data2 <- dplyr::sample_n(venn_data, 100) \%>\%
  rbind(data)
venn_net(data2, label.degree = 1,graph.layout = "fr")
venn_net(data2, edge.type = "hive", label.degree = 4)
}
}
