% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_plot.R
\name{venn_plot}
\alias{venn_plot}
\title{Intersection of datasets}
\usage{
venn_plot(
  data,
  set.color = c("#FF8748", "#5BAA56", "#4186B7", "#8679BE", "#B8BB5B"),
  name.size = 5,
  text.size = 4,
  bar.color = "Set1",
  type = 1,
  vratio = c(0.65, 0.35),
  ...
)
}
\arguments{
\item{data}{data.frame
column names must be two columns of data for gene and source}

\item{set.color}{color
Recommended color:
venn.color1 <- c("#b1d1e7", "#f0a4af", "#e5dfae", "#a8e1da", "#f8d068")
venn.color2 <- c("#d3e2b7", "#fbca50", "#c9d5d4", "#baa28a", "#4caecc")
venn.color3 <- c("#C1BCBF", "#B1D0A9", "#FFCCC3", "#FFD5AB", "#BCCBE5")
venn.color4 <- c("#e4c9b2", "#92C2DD", "#f49d98", "#fcd68f", "#629076")
venn.color5 <- c("#FF8748", "#5BAA56", "#B8BB5B", "#4186B7", "#8679BE")
venn.color6 <- c("#e6c09e", "#0d5888", "#cb8b3e", "#9cd6d6", "#dbcb09")
venn.color7 <- c("#E41A1C", "#1E90FF", "#FF8C00", "#4DAF4A", "#75cbdc")
venn.color8 <- c("#604192", "#139177", "#ed9e08", "#f56f08", "#4caecc")
venn.color9 <- c("#d37a20", "#dbcb09", "#3a9cbc", "#dd7208", "#a30019")}

\item{name.size}{name size}

\item{text.size}{text size}

\item{bar.color}{top column color}

\item{type}{1, 2
1:take the intersection of 2-4 data sets
2:take the intersection of >= 4 data sets}

\item{vratio}{The ratio of the histogram to the size of the matrix}

\item{...}{additional parameters}
}
\value{
figure
}
\description{
Intersection of datasets
}
\examples{
 \dontrun{
library(UpSetR)
data(venn_data, package = "TCMNP")
venn_plot(venn_data, type =1)
venn_plot(venn_data, type =2)
}

}
