% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_result.R
\name{venn_result}
\alias{venn_result}
\title{Intersection of datasets}
\usage{
venn_result(data)
}
\arguments{
\item{data}{data.frame
column names must be two columns of data for gene and source}
}
\value{
data.frame
}
\description{
Intersection of datasets
}
\examples{
 \dontrun{
data(venn_data, package = "TCMNP")
venn_result(venn_data)
}

}
