% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xfbdf}
\alias{xfbdf}
\title{Xuanfei Baidu granule (XFBD)}
\format{
An object of class \code{data.frame} with 5024 rows and 4 columns.
}
\source{
\url{http://www.nhc.gov.cn/cms-search/downFiles/a449a3e2e2c94d9a856d5faea2ff0f94.pdf}
}
\usage{
xfbdf
}
\description{
ephedra 6 g, bitter almond 15 g, raw gypsum 30 g,
raw almond 30 g, raw plaster 30 g,
Atractylodes macrocephala 10 g, patchouli 15 g,
artemisia grass 12 g, Polygonum cuspidatum 20 g,
verbena 30 g, dried Reed root 30 g,
Fructus thunbergii 15 g, tangerine red 15 g,
raw licorice 10 g.

A data frame with 5024 rows and 4 variables
}
\examples{
data("xfbdf", package = "TCMNP")
head(xfbdf)
}
\keyword{datasets}
